requirements.txt
llama_parse==0.4.2
nest_asyncio==1.5.8
langchain==0.1.13
openai==1.10.0
pinecone-client==3.2.2
pypdf2==3.0.1
python-dotenv==1.0.0
langchain_community==0.0.29